refactor(Checklist): GetChecklistsQuery
해당 퀴즈의 전체 체크리스트 조회를 위한 단일 책임 적용 및 클린 아키텍처 적용 UseCase
