+++ src/main/java/com/space/server/domain/checklist/application/port/in/GetChecklistsByQuizQuery.java
@@ -1,0 +1,9 @@
+package com.space.server.domain.checklist.application.port.in;
+
+import com.space.server.domain.checklist.domain.Checklist;
+
+import java.util.List;
+
+public interface GetChecklistsByQuizQuery {
+  List<Checklist> getChecklistsByQuiz(Long quizId);
+}
